services/image-processing/requirements.txt
fastapi
uvicorn
pillow  # Python Imaging Library for basic image processing tasks
